+++ test/single/as_void_test.dart
@@ -5,7 +5,7 @@ import 'package:rxdart_ext/rxdart_ext.dart';
 import 'package:test/expect.dart';
 import 'package:test/scaffolding.dart';
 
-import 'single_test_utils.dart';
+import 'utils.dart';
 
 void main() {
   group('Single.asVoid', () {