+++ test/single/async_expand_test.dart
@@ -2,7 +2,7 @@ import 'package:dart_either/dart_either.dart';
 import 'package:rxdart_ext/rxdart_ext.dart';
 import 'package:test/test.dart';
 
-import 'single_test_utils.dart';
+import 'utils.dart';
 
 void main() {
   group('asyncExpandSingle', () {